added click to move without pathfinding or collisions

+++ src/data/definitions/State.h
@@ -4,15 +4,15 @@
 class State
 {
 public:
-    inline int x() { return m_x; }
-    inline int y() { return m_y; }
+    inline float x() { return m_x; }
+    inline float y() { return m_y; }
 
-    inline void setX(int x) { m_x = x; }
-    inline void setY(int y) { m_y = y; }
+    inline void setX(float x) { m_x = x; }
+    inline void setY(float y) { m_y = y; }
 
 protected:
-    int m_x;
-    int m_y;
+    float m_x;
+    float m_y;
 };
 
 #endif //SFMLDEMO_STATE_H
